+++ src/web/configuration.ts
@@ -20,4 +20,18 @@ export default class ConfigurationService extends EventTarget {
     }
     this.dispatchEvent(new Event("configuration-changed"));
   }
+
+  UpdateLocalPrerelease(value: boolean): void {
+    if (this.configuration) {
+      this.configuration = { ...this.configuration, Prerelease: value };
+    }
+    this.dispatchEvent(new Event("configuration-changed"));
+  }
+
+  UpdateLocal(patch: Partial<Configuration>): void {
+    if (this.configuration) {
+      this.configuration = { ...this.configuration, ...patch };
+    }
+    this.dispatchEvent(new Event("configuration-changed"));
+  }
 }